Объединяет price и info
