+++ src/view/info.js
@@ -1,5 +1,22 @@
 import AbstractView from './abstract';
 
+const createPriceTemplate = (events) => {
+  let totalCoast = 0;
+  if (events.length) {
+    const totalPrice = events.reduce((acc, event) => acc + +event.price, 0);
+    const totalPriceOffers = events
+    .reduce((acc1, {offers, eventType}) => acc1 + offers
+    .filter(({checked, id}) => checked && eventType === id.toLowerCase())
+    .reduce((acc, {price}) => acc + +price, 0), 0);
+    totalCoast = totalPrice + totalPriceOffers;
+  }
+  return (
+    `<p class="trip-info__cost">
+      Total: &euro;&nbsp;<span class="trip-info__cost-value">${totalCoast}</span>
+    </p>`
+  );
+};
+
 const createRouteInfoTemplate = (events) => {
   return events.length > 3
     ? `${events[0].destination.city} &mdash; ... &mdash; ${events[events.length - 1].destination.city}`
@@ -25,6 +42,7 @@ const createInfoTemplate = (events) => {
         <h1 class="trip-info__title">${routeInfoTemplate}</h1>
         <p class="trip-info__dates">${dateStart}${dateFinish}</p>
       </div>
+      ${createPriceTemplate(events)}
     </section>`
   );
 };
